# docs/swagger.yaml
definitions:
...
  handler.auditLogResponse:
    properties:
      events:
        items:
          $ref: '#/definitions/handler.eventItem'
        type: array
      total:
        description: Total is the count of events matching the filters, ignoring limit/offset.
        example: 42
        type: integer
    type: object
...
  handler.errorResponse:
    properties:
      code:
        example: FORBIDDEN
        type: string
      docUrl:
        type: string
      error:
        example: dev or admin role required for mutations
        type: string
      hint:
        example: Ask an admin to grant you the dev role.
        type: string
      httpStatus:
        example: 403
        type: integer
    type: object
  handler.eventItem:
    properties:
      actor:
        example: asharma@armada.ai
        type: string
      changes:
        description: |-
          Changes is the pre-computed field-level diff. **Present ONLY for a clean
          single-entity update** (omitted otherwise via omitempty) — so its presence
          is the client's signal that a field diff is available; no need to inspect
          operations/before. Server-managed metadata (version/updatedAt/updatedBy/…)
          and DGraph UIDs are already excluded. Same data the UI's colored diff renders.
        items:
          $ref: '#/definitions/handler.fieldChange'
        type: array
      details:
        description: raw {operationName, query, variables, before}
        type: object
      eventCategory:
        description: data | management | auth
        example: data
        type: string
      id:
        example: 3d6bb15f-8c4c-45f0-8a6c-939b6f9cc512
        type: string
      operations:
        description: DGraph mutation fields, e.g. ["updateVeleroBackup"]
        items:
          type: string
        type: array
      resourceIds:
        description: orbIds touched
        items:
          type: string
        type: array
      resourceTypes:
        description: ConfigItem types touched, e.g. ["VeleroBackup"]
        items:
          type: string
        type: array
      timestamp:
        example: "2026-07-29T17:26:55Z"
        type: string
    type: object
  handler.exportRequest:
    properties:
      download:
        description: |-
          Download selects the mode. false (default) exports and publishes to OCI,
          discarding the zip on completion. true exports only and retains the zip
          for GET /api/v1/export/jobs/{jobId}/download.
        example: false
        type: boolean
      expectedContentHash:
        description: |-
          ExpectedContentHash is the optional guarded-Apply token: pass the
          `current.contentHash` from a prior POST /api/v1/export/preview and the
          export aborts with 409 if the data center's desired state changed since
          that preview (another writer edited intent in the review→Apply gap).
          Omit for last-writer-wins (default, back-compatible). This is the
          ETag/If-Match pattern applied to a whole subgraph — one hash rather than
          per-node version checks, so concurrent ADDS and DELETES are caught too.
        example: sha256:045b8a51a0aea59fa
        type: string
      orbId:
        description: OrbID identifies the data center to export, in "namespace:name"
          form. Required.
        example: alaska:dc-01
        type: string
    required:
    - orbId
    type: object
  handler.fieldChange:
    properties:
      after:
        description: new value (e.g. 15)
        type: object
      before:
        description: old value (e.g. 7)
        type: object
      field:
        example: retentionDays
        type: string
    type: object
...
  handler.restoreRequest:
    properties:
      backupId:
        description: BackupID is the UUID of the stored backup to restore from. Required.
        example: a1b2c3d4-0000-0000-0000-000000000000
        type: string
      confirmSchemaMismatch:
        description: |-
          ConfirmSchemaMismatch must be true to proceed when the backup's schema
          version differs from the running schema. false (default) makes such a
          restore fail fast rather than risk a mismatched load.
        example: false
        type: boolean
    required:
    - backupId
    type: object
...
  handler.triggerResponse:
    properties:
      id:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: API-first, graph-native source of truth for modular data centers.
  title: Orbital API
  version: 0.0.0-dev
paths:
  /api/v1/audit-log:
    get:
      description: |-
        Read-only, immutable audit trail of intent mutations, newest first.

        **Scope a query** by combining filters: `orbId` (repeatable, max 32) for a specific resource; `namespace` for a whole data center; `resource_type`/`operation_name` to narrow. To see everything under a server/cluster, fetch its subtree orbIds from the GraphQL Topology API and pass them as repeatable `orbId` params (there is no single "cluster" scope — a child mutation records the child's orbId, not the parent's).

        **Render a diff:** when an event is a clean single-entity update it carries a `changes` array (`[{field, before, after}]`) with metadata and DGraph UIDs already excluded — render it directly. When `changes` is absent (bulk add, create, or a multi-operation event), there is no field diff; fall back to showing `operations` + `resourceIds`. The raw `details` (with `before`/`variables`) is always included for callers that want it.

        Returns JSON by default; returns an HTML table fragment when the `HX-Request` header is present (used by orbital's own UI). See docs/api-cheatsheet.md § "Audit log".
      parameters:
      - description: Max results (default 100, max 500)
        in: query
        name: limit
        type: integer
      - description: Pagination offset
        in: query
        name: offset
        type: integer
      - collectionFormat: csv
        description: Filter by resource orbId (e.g. alaska-dot:GRTLY24). Repeatable
          (pass multiple), max 32 — matches events touching ANY of them.
        in: query
        items:
          type: string
        name: orbId
        type: array
      - description: Filter by namespace prefix (e.g. \
        in: query
        name: namespace
        type: string
      - description: RFC3339 lower bound (exclusive) on event timestamp
        in: query
        name: since
        type: string
      - description: RFC3339 upper bound (inclusive) on event timestamp
        in: query
        name: until
        type: string
      - description: Filter by resource ID
        in: query
        name: resource_id
        type: string
      - description: Filter by resource type (e.g. Server, DataCenter)
        in: query
        name: resource_type
        type: string
      - description: Filter to events containing this operation (exact, case-sensitive;
          stored form is verb-lowercased, e.g. updateVeleroBackup)
        in: query
        name: operation_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.auditLogResponse'
      summary: List audit events
      tags:
      - audit
...
  /api/v1/export:
    post:
      consumes:
      - application/json
      description: Triggers an async atomic export of the data center's configuration
        subgraph. When download=false (default) AND OCI is configured, the export
        chains into an OCI publish and the zip is discarded on completion. When download=true
        (or OCI is not configured), the export runs in download-only mode and the
        zip is retained on disk for the client to fetch via GET /api/v1/export/jobs/{jobId}/download.
        Returns 202 with a job ID immediately; poll GET /api/v1/export/jobs/{jobId}
        until terminal state. Returns 409 if an export or restore is already in progress,
        or if expectedContentHash is supplied and the desired state changed since
        that preview (guarded Apply).
      parameters:
      - description: Export request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.exportRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/handler.triggerResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Trigger subgraph export
      tags:
      - export
...
  /api/v1/restore:
    post:
      consumes:
      - application/json
      description: Restores DGraph blue from a stored backup. Blocked if any backup,
        export, or restore job is in progress.
      parameters:
      - description: Restore request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.restoreRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/handler.triggerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Trigger a DGraph restore
      tags:
      - backup
...
  /api/v1/users:
    get:
      description: Returns all users in the orbital user table. Admin only.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: List users
      tags:
      - users
...
  /graphql:
    post:
      consumes:
      - application/json
      description: 'POST: proxies GraphQL queries and mutations to DGraph. GET: serves
        the GraphiQL explorer UI.'
      parameters:
      - description: GraphQL request body
        example: '"{\"query\": \"{ queryDataCenter { id name } }\"}"'
        in: body
        name: body
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: GraphQL endpoint
      tags:
      - graphql
swagger: "2.0"
...